start to built object list

+++ src/vector.c
@@ -6,7 +6,7 @@
 /*   By: yizhang <yizhang@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/18 12:57:37 by yizhang       #+#    #+#                 */
-/*   Updated: 2023/10/25 14:30:01 by yizhang       ########   odam.nl         */
+/*   Updated: 2023/10/26 14:11:24 by yizhang       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ t_vec	sub(t_vec v1, t_vec v2)
 	return (new);
 }
 
-double	dot(t_vec v1, t_vec v2)
+float	dot(t_vec v1, t_vec v2)
 {
 	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
 }
@@ -45,14 +45,14 @@ t_vec	cross(t_vec v1, t_vec v2)
 	return (new);
 }
 
-double vec_len(t_vec v)
+float vec_len(t_vec v)
 {
     return sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
 }
 
 t_vec unit_vector(t_vec v)
 {
-    double len;
+    float len;
 	len = vec_len(v);
     t_vec ret;
 	ret = set_vec(v.x / len, v.y / len, v.z / len);
